LG-543 Retry failed Twilio Verify requests
**Why**: Sometimes, requests to the Twilio Verify service fail to
connect or time out. Twilio suggested that we retry the request to see
if it works the second time. We had implemented retries for the Twilio
REST API earlier. Now, we are adding it to the Verify service as well.

**How**:
- Use Faraday to catch and rescue timeouts and connection failures, and
raise our own custom `VerifyError`. Typhoeus does not raise exceptions,
and therefore does not support retries out of the box. Since the
`twilio-ruby` gem uses Faraday, and we rescue Faraday errors in
`TwilioService`, I thought we'd use Faraday here too for consistency.
- Use dependency injection to specify the connection class, which makes
testing easier.
- Include "Twilio Verify" in the custom error message to make it easier
to differentiate from timeout and failed connection errors raised by
the Twilio REST service.